+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Added `discard-invalid-images` ISP for removing corrupt images or annotations with no image attached
 - Added `batch-split` sub-command for splitting individual batches of annotations into subsets like train/test/val
 - Added `filter-metadata` ISP for filtering object detection
+- Restricted maximum characters per line in help output to 100 to avoid long help strings to become unreadable
 
 
 0.1.6 (2022-05-11)